Adopt clef rendering object partially
