+++ src/musicxml/clef.ts
@@ -1,4 +1,4 @@
-import { ClefAnnotation, ClefSign, ClefType, CLEF_SIGNS } from './enums';
+import { ClefSign, CLEF_SIGNS } from './enums';
 import { NamedElement } from '@/util';
 
 /**
@@ -27,54 +27,5 @@ export class Clef {
   /** Returns the octave change of the clef. Defaults to null. */
   getOctaveChange(): number | null {
     return this.element.first('clef-octave-change')?.content().int() ?? null;
-  }
-
-  /** Returns the clef type. Defaults to null. */
-  getClefType(): ClefType | null {
-    const sign = this.getSign();
-    const line = this.getLine();
-
-    if (sign === 'G') {
-      // with G line defaults to 2
-      // see https://www.w3.org/2021/06/musicxml40/musicxml-reference/elements/line/
-      if (line === 1) return 'french';
-      return 'treble';
-    }
-
-    if (sign === 'F') {
-      if (line === 5) return 'subbass';
-      if (line === 3) return 'baritone-f';
-      return 'bass';
-    }
-
-    if (sign === 'C') {
-      if (line === 5) return 'baritone-c';
-      if (line === 4) return 'tenor';
-      if (line === 2) return 'mezzo-soprano';
-      if (line === 1) return 'soprano';
-      return 'alto';
-    }
-
-    if (sign === 'percussion') {
-      return 'percussion';
-    }
-
-    if (sign === 'TAB') {
-      return 'tab';
-    }
-
-    return null;
-  }
-
-  /** Returns the clef annotation. Defaults to null. */
-  getAnnotation(): ClefAnnotation | null {
-    switch (this.getOctaveChange()) {
-      case 1:
-        return '8va';
-      case -1:
-        return '8vb';
-      default:
-        return null;
-    }
   }
 }
