+++ src/rendering/clef.ts
@@ -1,4 +1,5 @@
 import * as musicxml from '@/musicxml';
+import { ClefType } from './enums';
 
 /** A musical symbol used to indicate which notes are represented by the lines and spaces on a stave. */
 export class Clef {
@@ -20,5 +21,45 @@ export class Clef {
     const octaveChange = clef.getOctaveChange();
 
     return new Clef(sign, line, octaveChange);
+  }
+
+  isEqual(other: Clef): boolean {
+    return this.sign === other.sign && this.line === other.line && this.octaveChange === other.octaveChange;
+  }
+
+  getType(): ClefType | null {
+    const sign = this.sign;
+    const line = this.line;
+
+    if (sign === 'G') {
+      // with G line defaults to 2
+      // see https://www.w3.org/2021/06/musicxml40/musicxml-reference/elements/line/
+      if (line === 1) return 'french';
+      return 'treble';
+    }
+
+    if (sign === 'F') {
+      if (line === 5) return 'subbass';
+      if (line === 3) return 'baritone-f';
+      return 'bass';
+    }
+
+    if (sign === 'C') {
+      if (line === 5) return 'baritone-c';
+      if (line === 4) return 'tenor';
+      if (line === 2) return 'mezzo-soprano';
+      if (line === 1) return 'soprano';
+      return 'alto';
+    }
+
+    if (sign === 'percussion') {
+      return 'percussion';
+    }
+
+    if (sign === 'TAB') {
+      return 'tab';
+    }
+
+    return null;
   }
 }
