+++ tests/unit/musicxml/clef.test.ts
@@ -69,55 +69,4 @@ describe(Clef, () => {
       expect(clef.getOctaveChange()).toBeNull();
     });
   });
-
-  describe('getClefType', () => {
-    it.each([
-      { sign: 'G', line: 1, clefType: 'french' },
-      { sign: 'G', line: undefined, clefType: 'treble' },
-      { sign: 'F', line: 5, clefType: 'subbass' },
-      { sign: 'F', line: 3, clefType: 'baritone-f' },
-      { sign: 'F', line: undefined, clefType: 'bass' },
-      { sign: 'C', line: 5, clefType: 'baritone-c' },
-      { sign: 'C', line: 4, clefType: 'tenor' },
-      { sign: 'C', line: 2, clefType: 'mezzo-soprano' },
-      { sign: 'C', line: 1, clefType: 'soprano' },
-      { sign: 'C', line: undefined, clefType: 'alto' },
-      { sign: 'percussion', line: undefined, clefType: 'percussion' },
-      { sign: 'TAB', line: undefined, clefType: 'tab' },
-      { sign: 'foo', line: 1, clefType: null },
-    ])(`returns the clef type for sign '$sign' and line $line`, (t) => {
-      const node = xml.clef({ sign: xml.sign({ value: t.sign }), line: xml.line({ value: t.line }) });
-      const clef = new Clef(node);
-      expect(clef.getClefType()).toBe(t.clefType);
-    });
-
-    it('defaults to null when the clef sign is invalid', () => {
-      const node = xml.clef({ sign: xml.sign({ value: 'foo' }) });
-      const clef = new Clef(node);
-      expect(clef.getClefType()).toBeNull();
-    });
-  });
-
-  describe('getAnnotation', () => {
-    it.each([
-      { octaveChange: 1, annotation: '8va' },
-      { octaveChange: -1, annotation: '8vb' },
-    ])('returns the annotation of the clef: octaveChange $octaveChange', (t) => {
-      const node = xml.clef({ clefOctaveChange: xml.clefOctaveChange({ value: t.octaveChange }) });
-      const clef = new Clef(node);
-      expect(clef.getAnnotation()).toBe(t.annotation);
-    });
-
-    it('returns null when the octave change is invalid', () => {
-      const node = xml.clef({ clefOctaveChange: xml.clefOctaveChange({ value: NaN }) });
-      const clef = new Clef(node);
-      expect(clef.getAnnotation()).toBeNull();
-    });
-
-    it('returns null when the octave change is missing', () => {
-      const node = xml.clef();
-      const clef = new Clef(node);
-      expect(clef.getAnnotation()).toBeNull();
-    });
-  });
 });
